Split canonical CBOR helpers
